CAST values + corrections

+++ aesh_suivi.sql
@@ -5,7 +5,7 @@ SET group_id = (SELECT user_info.groupe FROM login_session join user_info on use
 
 SELECT 'redirect' AS component,
         'etablissement.sql?restriction' AS link
-        WHERE $group_id::int<'2';
+        WHERE $group_id<'2';
 
 --Menu
 SELECT 'dynamic' AS component, sqlpage.read_file_as_text('menu.json') AS properties;
@@ -19,13 +19,13 @@ select
     'Modifier' as title,
     'pencil' as icon,
     'orange' as outline,
-        $group_id::int<3 as disabled,
+        $group_id<3 as disabled,
     'aesh_edit.sql?id='||$id||'&username='||username as link FROM aesh WHERE aesh.id = $id;
 select 
     'Planning' as title,
     'clock-share' as icon,
     'red' as outline,
-        $group_id::int<3 as disabled,
+        $group_id<3 as disabled,
     'upload_edt_form.sql?id='||$id as link FROM aesh where aesh.id=$id;
    
 -- écrire le nom de l'AESH dans le titre de la page
@@ -36,7 +36,7 @@ SELECT
     aesh_name||' '||aesh_firstname as description, 'orange' as color, 1 as active
      FROM aesh WHERE aesh.id = $id;
 SELECT 
-    CASE WHEN $group_id::int>2 
+    CASE WHEN $group_id>2 
     THEN    tel_aesh 
     ELSE 'numéro masqué'
     END as title,
