+++ etab_classes.sql
@@ -183,7 +183,7 @@ THEN
 ](notification.sql?id='||eleve.id||'&tab=Examen "Aménagements d''examen")' 
 ELSE ''
 END as Actions,
-CASE WHEN $group_id::int>2 THEN
+CASE WHEN $group_id>2 THEN
 '[
     ![](./icons/pencil.svg)
 ](eleve_edit.sql?id='||eleve.id||')'
